Phone display (#14)
* fixed login page display

* navbar now dynamically changes size

* navbar on phone now has a dropdown menu

* added highlights to dropdown menu

* fixed some issues with navbar highlights

* dropdown menu now closes when clicked out of on mobile

* changed weeklong info to automatically format text files

* events page mobile display supported

* event statistics mobile display support

* fixed navbar profile/logout buttons not showing up

* removed tables and replaced with div tags

* fixed zero kills display

* fixed username overflow display

+++ classes/user.php
@@ -93,6 +93,9 @@ class User extends Password{
 
 	// this will remove a user that is signed from an event
 	public function leave_event($event){
+		if(!$this->is_in_event($event)){
+			return false;
+		}
 		try{
 			$stmt = $this->_db->prepare('DELETE FROM '.$event.' WHERE user_id=:user_id');
 			$stmt->execute(array(
